refactor: extract GitHub contribution into contribution.ts
contribute, closeContribution, both PR message generators and
getSpecificationsFilesList move out as functions; the class keeps
delegates (route + poller entry points). The write-only fields
differentFilename/notBackwardCompatible and the unreachable
backward-compatibility branch are gone (TODO comment documents the
never-implemented check).
