+++ backend/src/specification/index.ts
@@ -7,6 +7,7 @@ export * from './modbusValues.js'
 export * from './specMessages.js'
 export * from './specDiff.js'
 export * from './specValidator.js'
+export * from './contribution.js'
 export * from './m2mgithub.js'
 export * from './configspec.js'
 export * from './migrator.js'
