+++ backend/src/specification/m2mspecification.ts
@@ -1,34 +1,17 @@
 import { IspecificationValidator, IvalidateIdentificationResult } from './ispecificationvalidator.js'
-import { join } from 'path'
-import * as fs from 'fs'
 import Debug from 'debug'
 import { Idata, IfileSpecification } from './ifilespecification.js'
-import { M2mGitHub } from './m2mgithub.js'
-import {
-  Imessage,
-  MessageTypes,
-  MessageCategories,
-  VariableTargetParameters,
-  getParameterType,
-  validateTranslation,
-  ModbusRegisterType,
-} from '../shared/specification/index.js'
+import { Imessage, VariableTargetParameters } from '../shared/specification/index.js'
 import {
   Ispecification,
   IbaseSpecification,
   SpecificationStatus,
   getSpecificationI18nName,
   ImodbusSpecification,
-  getSpecificationI18nEntityName,
   SpecificationFileUsage,
-  FileLocation,
   IdentifiedStates,
-  ISpecificationText,
   ImodbusEntity,
   Inumber,
-  IminMax,
-  Iselect,
-  Itext,
 } from '../shared/specification/index.js'
 import { ConfigSpecification } from './configspec.js'
 import { LogLevelEnum, Logger } from './log.js'
@@ -39,9 +22,9 @@ import {
   copyModbusDataToEntity,
   emptyModbusValues,
   fileToModbusSpecification,
-  setIdentifiedByEntities,
 } from './modbusValues.js'
 import { getMessageString, messages2Text } from './specMessages.js'
+import { closeContribution, contribute, getSpecificationsFilesList } from './contribution.js'
 import { compareSpecifications, isEqualValue } from './specDiff.js'
 import { validateBaseSpecification, validateFiles, validateSpecification, validateUniqueName } from './specValidator.js'
 import { Observable, Subject } from 'rxjs'
@@ -83,154 +66,17 @@ export class M2mSpecification implements IspecificationValidator {
     return messages2Text(spec, msgs)
   }
   async contribute(note: string | undefined): Promise<number> {
-    const language = ConfigSpecification.mqttdiscoverylanguage
-    let messages: Imessage[] = []
-
-    if (language == undefined)
-      messages.push({ type: MessageTypes.noMqttDiscoveryLanguage, category: MessageCategories.configuration })
-    else messages = this.validate(language)
-    const errors: string = M2mSpecification.messages2Text(this.settings as IbaseSpecification, messages)
-
-    if (errors.length > 0) {
-      throw new Error('Validation failed with errors: ' + errors)
-    }
-
-    if (errors.length == 0 && messages.length > 0 && (!note || note.length == 0))
-      throw new Error('Validation failed with warning, but no note text available')
-    const fileList = this.getSpecificationsFilesList(ConfigSpecification.getLocalDir())
-    const spec = this.settings as IbaseSpecification
-    let title = ''
-    let message = ''
-    switch (spec.status) {
-      case SpecificationStatus.added: {
-        title = 'Add specification '
-        message = this.generateAddedContributionMessage()
-        break
-      }
-      case SpecificationStatus.cloned: {
-        title = 'Update specification '
-        const pub = (spec as unknown as { publicSpecification?: IfileSpecification }).publicSpecification
-        message = this.generateClonedContributionMessage(note, pub)
-        break
-      }
-    }
-    title = title + getSpecificationI18nName(spec, language!)
-    if (!ConfigSpecification.githubPersonalToken || !ConfigSpecification.githubPersonalToken.length) {
-      throw new Error('Github connection is not configured. Set Github Personal Acces Token in configuration UI first')
-    }
-    const github = new M2mGitHub(ConfigSpecification.githubPersonalToken, ConfigSpecification.getPublicDir())
-    try {
-      await github.init()
-      await github.commitFiles(ConfigSpecification.getLocalDir(), spec.filename, fileList, title, message)
-      const issue = await github.createPullrequest(title, message, spec.filename)
-      new ConfigSpecification().changeContributionStatus(
-        (this.settings as IbaseSpecification).filename,
-        SpecificationStatus.contributed,
-        issue
-      )
-      return issue
-    } catch (e: unknown) {
-      if (spec.status == SpecificationStatus.contributed)
-        new ConfigSpecification().changeContributionStatus(spec.filename, SpecificationStatus.added)
-      try {
-        await github.deleteSpecBranch(spec.filename)
-      } catch (e1: unknown) {
-        log.log(LogLevelEnum.error, 'delete branch: ' + (e1 as Error).message)
-      }
-      throw e
-    }
+    return contribute(this.settings as Ispecification, note)
   }
 
-  private generateAddedContributionMessage(): string {
-    // First contribution:
-    // Name of Specification(en)
-    const spec = this.settings as ImodbusSpecification
-    let message = `First contribution of ${getSpecificationI18nName(spec, 'en')}(${spec.filename}) \nEntities:\n`
-    message = `${message}Languages: `
-    spec.i18n.forEach((l) => {
-      message = `${message} ${l.lang} `
-    })
-    message = `${message}\nEntities:\n`
-    spec.entities.forEach((ent) => {
-      message = `${message}\t${getSpecificationI18nEntityName(spec, 'en', ent.id)}\n`
-    })
-    message = `${message}\nImages:\n`
-    spec.files.forEach((file) => {
-      if (file.usage == SpecificationFileUsage.img) message = `${message}\t ${file.url}\n`
-    })
-    message = `${message}\nDocumentation:\n`
-    spec.files.forEach((file) => {
-      if (file.usage == SpecificationFileUsage.documentation) message = `${message}\t ${file.url}\n`
-    })
-    return message
-  }
-
-  private generateClonedContributionMessage(note: string | undefined, publicSpecification: IfileSpecification | undefined): string {
-    let rcmessage = ''
-    this.notBackwardCompatible = false
-    this.differentFilename = false
-    if (publicSpecification) {
-      rcmessage = rcmessage + 'Changes:\n'
-      const messages = this.isEqual(publicSpecification)
-      messages.forEach((message) => {
-        rcmessage = rcmessage + M2mSpecification.getMessageString(this.settings as IbaseSpecification, message) + '\n'
-      })
-      // TODO Check backward compatibility
-      if (this.notBackwardCompatible) {
-        rcmessage = rcmessage + '\n!!! There are changes which are not backward compatible !!'
-        if (note == undefined) throw new Error('There are changes which are not backward compatible')
-      }
-
-      if (note != undefined) rcmessage = rcmessage + '\n' + note
-    }
-    return rcmessage
-  }
   static getMessageString(spec: IbaseSpecification, message: Imessage): string {
     return getMessageString(spec, message)
   }
-  private static throwCloseContributionError(msg: string): never {
-    log.log(LogLevelEnum.error, msg)
-    const e = new Error(msg) as Error & { step?: string }
-    e.step = 'closeContribution'
-    throw e
-  }
-  static async closeContribution(spec: IfileSpecification): Promise<IpullRequest> {
-    if (undefined == ConfigSpecification.githubPersonalToken) {
-      this.throwCloseContributionError(
-        'No Github Personal Access Token configured. Unable to close contribution ' + spec.filename
-      )
-    }
-    if (spec.pullNumber == undefined) {
-      this.throwCloseContributionError('No Pull Number in specification. Unable to close contribution ' + spec.filename)
-    }
-    const gh = new M2mGitHub(ConfigSpecification.githubPersonalToken!, join(ConfigSpecification.getPublicDir()))
-    try {
-      await gh.init()
-      const pullStatus = await gh.getPullRequest(spec.pullNumber!)
-      const cspec = new ConfigSpecification()
-      if (pullStatus.merged) {
-        cspec.changeContributionStatus(spec.filename, SpecificationStatus.published, undefined)
-      } else if (pullStatus.closed_at != null) {
-        cspec.changeContributionStatus(spec.filename, SpecificationStatus.added, undefined)
-      }
-      spec = ConfigSpecification.getSpecificationByFilename(spec.filename)!
-      if (spec.status != SpecificationStatus.contributed) gh.deleteSpecBranch(spec.filename)
-      gh.fetchPublicFiles()
-      return { merged: pullStatus.merged, closed: pullStatus.closed_at != null, pullNumber: spec.pullNumber! }
-    } catch (e: unknown) {
-      const msg = e instanceof Error ? e.message : String(e)
-      this.throwCloseContributionError('closeContribution: ' + msg)
-    }
+  static closeContribution(spec: IfileSpecification): Promise<IpullRequest> {
+    return closeContribution(spec)
   }
   getSpecificationsFilesList(localDir: string): string[] {
-    const files: string[] = []
-    const spec = this.settings as IbaseSpecification
-    spec.files.forEach((file) => {
-      const filePath = file.url.replace(/^\//g, '')
-      if (file.fileLocation == FileLocation.Local && fs.existsSync(join(localDir, filePath))) files.push(filePath)
-    })
-    files.push(join('specifications', spec.filename + '.json'))
-    return files
+    return getSpecificationsFilesList(this.settings as IbaseSpecification, localDir)
   }
 
   validate(language: string): Imessage[] {
